Adding ARN
These changes were made in an effort to add the principal as the callerARN to the policy simulator. However when this was done, the resource policy no longer enforced the users access. This was because the callerARN was used without a ResourceOwner so the callARN was used as the resourceOwner. This is the expected behavior of the policy simulator according to AWS. If the callerARN is in the same account as the resourceARN then the resource policy is not used.

solution 1: is to include a random garbage resourceOwner. This will use the resource policy.
solution 2: populate the Principal field of the resource policy with the specific accounts, users that can access this resource